refactor: move taproot utils file
