+++ ts_src/payments/taprootutils.ts
@@ -1,20 +1,23 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
-exports.tapTweakHash = exports.tapLeafHash = exports.findScriptPath = exports.toHashTree = exports.rootHashFromPath = void 0;
-const buffer_1 = require('buffer');
-const bcrypto = require('./crypto');
-// todo: use varuint-bitcoin??
-const varuint = require('bip174/src/lib/converter/varint');
-// todo: !!!Temp, to be replaced. Only works because bip32 has it as dependecy. Linting will fail.
-// const ecc = require('tiny-secp256k1');
+import { Buffer as NBuffer } from 'buffer';
+import * as bcrypto from '../crypto';
+
+import { varuint } from '../bufferutils';
+import { TaprootLeaf } from '../types';
+
 const LEAF_VERSION_TAPSCRIPT = 0xc0;
 const TAP_LEAF_TAG = 'TapLeaf';
 const TAP_BRANCH_TAG = 'TapBranch';
 const TAP_TWEAK_TAG = 'TapTweak';
-function rootHashFromPath(controlBlock, tapLeafMsg) {
+
+export function rootHashFromPath(
+  controlBlock: Buffer,
+  tapLeafMsg: Buffer,
+): Buffer {
   const k = [tapLeafMsg];
   const e = [];
+
   const m = (controlBlock.length - 33) / 32;
+
   for (let j = 0; j < m; j++) {
     e[j] = controlBlock.slice(33 + 32 * j, 65 + 32 * j);
     if (k[j].compare(e[j]) < 0) {
@@ -23,10 +26,17 @@ function rootHashFromPath(controlBlock, tapLeafMsg) {
       k[j + 1] = tapBranchHash(e[j], k[j]);
     }
   }
+
   return k[m];
 }
-exports.rootHashFromPath = rootHashFromPath;
-function toHashTree(scripts) {
+
+export interface HashTree {
+  hash: Buffer;
+  left?: HashTree;
+  right?: HashTree;
+}
+
+export function toHashTree(scripts: TaprootLeaf[]): HashTree {
   if (scripts.length === 1) {
     const script = scripts[0];
     if (Array.isArray(script)) {
@@ -35,6 +45,7 @@ function toHashTree(scripts) {
     script.version = script.version || LEAF_VERSION_TAPSCRIPT;
     if ((script.version & 1) !== 0)
       throw new TypeError('Invalid script version');
+
     return {
       hash: tapLeafHash(script.output, script.version),
     };
@@ -43,8 +54,10 @@ function toHashTree(scripts) {
   const half = Math.trunc(scripts.length / 2);
   const left = toHashTree(scripts.slice(0, half));
   const right = toHashTree(scripts.slice(half));
+
   let leftHash = left.hash;
   let rightHash = right.hash;
+
   if (leftHash.compare(rightHash) === 1)
     [leftHash, rightHash] = [rightHash, leftHash];
   return {
@@ -53,47 +66,47 @@ function toHashTree(scripts) {
     right,
   };
 }
-exports.toHashTree = toHashTree;
-function findScriptPath(node, hash) {
+
+export function findScriptPath(node: HashTree, hash: Buffer): Buffer[] {
   if (node.left) {
     if (node.left.hash.equals(hash)) return node.right ? [node.right.hash] : [];
     const leftPath = findScriptPath(node.left, hash);
     if (leftPath.length)
       return node.right ? [node.right.hash].concat(leftPath) : leftPath;
   }
+
   if (node.right) {
     if (node.right.hash.equals(hash)) return node.left ? [node.left.hash] : [];
     const rightPath = findScriptPath(node.right, hash);
     if (rightPath.length)
       return node.left ? [node.left.hash].concat(rightPath) : rightPath;
   }
+
   return [];
 }
-exports.findScriptPath = findScriptPath;
-function tapLeafHash(script, version) {
+
+export function tapLeafHash(script: Buffer, version?: number): Buffer {
   version = version || LEAF_VERSION_TAPSCRIPT;
   return bcrypto.taggedHash(
     TAP_LEAF_TAG,
-    buffer_1.Buffer.concat([
-      buffer_1.Buffer.from([version]),
-      serializeScript(script),
-    ]),
+    NBuffer.concat([NBuffer.from([version]), serializeScript(script)]),
   );
 }
-exports.tapLeafHash = tapLeafHash;
-function tapBranchHash(a, b) {
-  return bcrypto.taggedHash(TAP_BRANCH_TAG, buffer_1.Buffer.concat([a, b]));
+
+function tapBranchHash(a: Buffer, b: Buffer): Buffer {
+  return bcrypto.taggedHash(TAP_BRANCH_TAG, NBuffer.concat([a, b]));
 }
-function tapTweakHash(pubKey, h) {
+
+export function tapTweakHash(pubKey: Buffer, h: Buffer | undefined): Buffer {
   return bcrypto.taggedHash(
     TAP_TWEAK_TAG,
-    buffer_1.Buffer.concat(h ? [pubKey, h] : [pubKey]),
+    NBuffer.concat(h ? [pubKey, h] : [pubKey]),
   );
 }
-exports.tapTweakHash = tapTweakHash;
-function serializeScript(s) {
+
+function serializeScript(s: Buffer): Buffer {
   const varintLen = varuint.encodingLength(s.length);
-  const buffer = buffer_1.Buffer.allocUnsafe(varintLen); // better
+  const buffer = NBuffer.allocUnsafe(varintLen); // better
   varuint.encode(s.length, buffer);
-  return buffer_1.Buffer.concat([buffer, s]);
+  return NBuffer.concat([buffer, s]);
 }
